login_gui files added for twitter
login_gui2.js file follows the same design pattern as we used for the
other social website, such as facebook, instagram and so forth.
However, due to the chrome content security limitation, the
login_gui.js file seems to be a better solution to tackle the login
function for twitter in our extension.